django/requirements.txt
celery==4.0.2
digest==1.0.2
Django==1.11
feedparser==5.2.1
djangorestframework==3.6.3
django-webpack-loader==0.4.1
django-nose==1.4.4

